AccessDB

Documentation

getArtistID :: String -> IO Int

The getArtistID function takes a string (The artist Name) and looks for any artists in the database with this name (case insensitive). If there are any it returns the id, otherwise -1

getAlbumInfo :: Int -> IO [(Int, String)]

The getAlbumInfo function takes an Int (The artist ID) and returns All of the ID's and Name's of the albums related to that artist. These are in the form of a List of Pairs

getTrackInfo :: Int -> IO [(String, String)]

The getTrackInfo function takes an Int (The album ID) and returns the Names and Preview Urls of all the Tracks related to that album in the form of a list of pairs

getImageURL :: Int -> IO String

The getImageUrl function takes an Int (the album ID) and returns that albums artwork. If no artwork is available, then returns a default image