GenerateHTML

Documentation

createPage :: String -> IO String

The create page function takes a string argument consistsing of an artists name. It checks if the artist is in the database and if not attempts to add them If this is successful, the artist page is build, if not the NoArtist page is returned (found in the HTML folder)