SaveToDB

Documentation

checkArtistInDB :: String -> Connection -> IO Bool

The checkArtistInDB function takes an artist name as a string and a Connection object. It then returns a boolean denoting weather of not the artist is contained within the artists table

addArtistToDB :: FullArtist -> Connection -> IO ()

The addArtistToDB function Takes a FullArtist Object and a connection object. It adds the FullArtist to the database, and calls teh addGenreToDB function to deal with the artists genres

addAlbumToDB :: [Album] -> String -> Connection -> IO ()

The addAlbumToDB function takes a list of albums, the name of the artist that the albums belong to, and a connection object. It firstly checks if any of the albums are duplicates (as there are many in the spotify api). For the albums for which this is not the case, it adds their information to the albums table, and their artwork to the images table.

addTracksToDB :: [Track] -> String -> Connection -> IO ()

The addTracksToDB function takes a list of Tracks a spotiyf album ID (String) and a connection object. It first checks that the album ID is in the database (due to removing the doubles), and if it is adds all of the tracks in the list into the tracks table.