SpotifyDataTypes

Documentation

data Info

This is the type used to store all the artists returned from a search

Constructors

Info

Fields

artists :: [BasicArtist]

data BasicArtist

The Basic Artist is the information for artists retured from a search query. This is then used to find the full artist object

Constructors

BasicArtist

Fields

href' :: String

name' :: String

popularity' :: String

data FullArtist

The full artist contains all the information about the artist that is returned from a query to the Spotify Artist Rest API

Constructors

FullArtist

Fields

followers :: Int

genres :: [String]

identifier :: String

name :: String

popularity :: Double

data Albums

The Albums type stores a list of all albums related to an artist

Constructors

Albums

Fields

albums :: [Album]

data Album

The Album type stores information on Artist albums, most impotantly their ID used to identify the tracks it contains

Constructors

Album

Fields

albumID :: String

albumImages :: [Image]

albumName :: String

data Image

The Image type contains the link and dimentions of album artwork

Constructors

Image

Fields

height :: Int

imageurl :: String

width :: Int

data Tracks

The Tracks type contains a list of all tracks from an Album

Constructors

Tracks

Fields

tracks :: [Track]

data Track

The Track type contains metadata about songs available on spotify. They also contain the preview URL used in the project to listen to samples of the ALbums

Constructors

Track

Fields

duration :: Int

explicit :: Bool

trackId :: String

trackName :: String

preview :: String

trackNum :: Int